buffer_utils: Moved the buffer <-> integer read/write code into a new header
